refactor : Deck의 인터페이스화
